Change image to RGB after loading
This fixes a bug where the script would crash if a PNG was loaded with
RGBA values. We're not expecting alpha values for a cross-stitch, so
we remove them. Might be worth understanding what this conversion
actually does so we can be sure that it does something sensible when
an image has an alpha channel.